+++ dataans/src-tauri/src/dataans/service/mod.rs
@@ -1,0 +1,2 @@
+pub mod note;
+pub mod space;
